+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "autoprefixer": "^10.4.20",
     "clsx": "^2.1.1",
+    "dayjs": "^1.11.13",
     "framer-motion": "^11.13.1",
     "postcss": "^8.4.49",
     "react": "^18.3.1",
