+++ src/components/List/type/index.ts
@@ -1,0 +1,7 @@
+export type ListItem = {
+  id?: number;
+  title: React.ReactNode;
+  subtitle?: React.ReactNode;
+  avatar?: string;
+  content?: React.ReactNode;
+};
